+++ Toggly.FeatureManagement.Nuxt/tests/hosts/fixture/server/api/definitions/[...path].get.ts
@@ -1,0 +1,5 @@
+export default defineEventHandler(event => ({
+  Enabled: getQuery(event).u !== 'bob',
+  Disabled: false,
+  Targeted: getQuery(event).u === 'alice',
+}))